chore: lee -> starCount와 constellationCount를 int형으로 변경

+++ unibirth-BE/src/main/java/com/ssafy/unibirth/star/service/StarService.java
@@ -32,7 +32,7 @@ public class StarService {
     @Transactional
     public CreateStarResDto create(CreateStarReqDto dto) {
         Member member = memberService.getCurrentMember();
-        Long starCount = member.plusStarCount();
+        int starCount = member.plusStarCount();
 
         Long constellationId = dto.getConstellationId();
         checkCompletion(constellationId);
